Relicense under LGPL and add contact information.
Licensing a standalone software (and even less a plugin) under
AGPL doesn’t make sense as it is meant just for software used as
a service. Moreover, I don’t it is even lawful to license a
plugin of non-GPL software with a viral license. Therefore, I am
relicensing this with LGPL, which is in my opinion the closest
legal option to the original authors intent.